// src/services/auth/enums.ts
export enum AuthApiErrorCode {
  LoginFailed = 'AUTH_LOGIN_FAILED',
  InvalidInvitation = 'AUTH_INVALID_INVITAION',
  InvalidToken = 'AUTH_INVALID_TOKEN',
  AlreadyExist = 'AUTH_ALREADY_EXIST',
  PasswordsMismatch = 'AUTH_PASSWORDS_MISMATCH',
  WrongPassword = 'AUTH_WRONG_PASSWORD',
  SamePassword = 'AUTH_SAME_PASSWORD',
  Blocked = 'AUTH_BLOCKED',
}